backend/requirements.txt
django>=4.2,<5.0
djangorestframework
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
django-cors-headers
django-filter
drf-spectacular
